requirements.txt
requests
beautifulsoup4
pillow
google-genai

python-dotenv
